+++ hotness/helpers.py
@@ -193,7 +193,7 @@ def upstream_cmp(v1, v2):
     return 0
 
 
-__rc_upstream_regex = re.compile("(.*?)\.?(-?(rc|pre|beta|alpha|dev)([0-9]*))", re.I)
+__rc_upstream_regex = re.compile(r"(.*?)\.?(-?(rc|pre|beta|alpha|dev)([0-9]*))", re.I)
 __rc_release_regex = re.compile(r"0\.[0-9]+\.(rc|pre|beta|alpha|dev)([0-9]*)", re.I)